requirements.txt
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.2.2
blessed==1.19.1
boto3==1.24.35
botocore==1.27.35
braintree==4.15.2
cached-property==1.5.2
cement==2.8.2
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
cryptography==37.0.4
dj-database-url==0.5.0
Django==4.0.4
django-admin-interface==0.19.2
django-ckeditor==6.5.0
django-colorfield==0.7.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-js-asset==2.0.0
django-nested-admin==3.4.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
jsonschema==3.2.0
numpy==1.22.3
pandas==1.4.2
paramiko==2.11.0
pathspec==0.9.0
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyotp==2.6.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-monkey-business==1.0.0
pytz==2022.1
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.6.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
stripe==2.63.0
termcolor==1.1.0
texttable==1.6.4
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.1.9
websocket-client==0.59.0
whitenoise==6.0.0
XlsxWriter==3.0.3
